requirements-xpu.txt
--extra-index-url https://download.pytorch.org/whl/xpu

numpy
torchtune
torchao
tqdm
traitlets
traittypes
transformers
tokenizers
ipykernel
einops
accelerate
bitsandbytes
vector-quantize-pytorch
modelscope
soundfile
torchcodec
torchaudio
gradio